requirements.txt
streamlit
pandas
plotly
python-dotenv
alpaca-py